+++ src/svg-renderer.js
@@ -1,48 +1,33 @@
 import * as PIXI from 'pixi.js';
 
 /**
- * Promised-based SVG to image converter
- * @param   {string|object}  svg  SVG string or DOM element to be converted
- * @return  {object}  Promise resolving to the image
+ * SVG to PIXI texture converter
+ * @param {string|object} svg - SVG string or DOM element to be converted
+ * @return {object} Rendered texture
  */
-export const svgToImg = (
+const renderSvg = (
   svg,
   { width = null, height = null, background = null } = {}
-) =>
-  new Promise((resolve, reject) => {
-    const image = new Image();
-    if (width !== null && height !== null) {
-      image.width = width;
-      image.height = height;
-    }
+) => {
+  let svgStr =
+    typeof svg === 'string' || svg instanceof String
+      ? svg
+      : new XMLSerializer().serializeToString(svg);
 
-    // serialize svg element into a string if needed
-    let svgStr =
-      typeof svg === 'string' || svg instanceof String
-        ? svg
-        : new XMLSerializer().serializeToString(svg);
+  const widthAttr = width && `width="${width}"`;
+  const heightAttr = height && `height="${height}"`;
+  const attrs = [widthAttr, heightAttr].filter(s => s).join(' ');
 
-    const style = background ? `style="background: ${background}"` : '';
-    svgStr = `${svgStr.slice(0, 5)} ${style} ${svgStr.slice(5)}`;
+  const backgroundStyle = background && `background: ${background}`;
+  const styles = [backgroundStyle].filter(s => s);
+  const style = `style="${styles.join('; ')}"`;
 
-    // convert SVG string to base64
-    const image64 = `data:image/svg+xml;base64,${btoa(svgStr)}`;
+  svgStr = `${svgStr.slice(0, 5)} ${attrs} ${style} ${svgStr.slice(5)}`;
 
-    image.onload = () => {
-      resolve(image);
-    };
-    image.onerror = error => {
-      reject(error);
-    };
-
-    image.src = image64;
-  });
-
-const renderImage = image => PIXI.Texture.from(image);
+  return PIXI.Texture.from(svgStr);
+};
 
 const createSvgRenderer = options => sources =>
-  Promise.all(
-    sources.map(src => svgToImg(src, options).then(image => renderImage(image)))
-  );
+  Promise.all(sources.map(src => renderSvg(src, options)));
 
 export default createSvgRenderer;